pyproject.toml
[tool.poetry]
name = "profyle"
version = "0.2.1-beta1"
description = "Profyle, a development tool for analysing and managing python traces"
authors = ["Carlos Valdivia <vpcarlos97@gmail.com>"]
maintainers = [
    "Carlos Valdivia <vpcarlos97@gmail.com>"
]
repository = "https://github.com/vpcarlos/profyle"
keywords = ["cprofile","devtool", "perffeto", "profyle", "viztracer","fastapi","flask"]
classifiers = [
    "Framework :: FastAPI",
    "Framework :: Flask",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python",
    "Topic :: Internet",
    "Topic :: Software Development :: Debuggers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development",
    "Typing :: Typed",
    "Environment :: Web Environment",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
profyle = "profyle.main:app"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
jinja2 = "^3.1.2"
uvicorn = ">=0.20.0"
typer = {extras = ["all"], version = "^0.7.0"}
viztracer = "^0.16.0"
fastapi = ">=0.70.0"
pydantic-settings = "^2.0.3"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"
pytest = "^7.2.1"
httpx = "^0.23.3"
flask = "^2.0.0"
pytest-asyncio = "^0.21.1"
ruff = "^0.1.5"


[tool.poetry.group.flask.dependencies]
flask = "^2.0.0"


[tool.poetry.group.django.dependencies]
django = "^4.2.7"

[tool.ruff]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP",
]

ignore = [
    # Avoid trying to fix flake8-bugbear (`B`) violations.
    "B",
    "E731",
]

fix = true
unfixable = []

line-length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
